% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{evaluate_model_ptr}
\alias{evaluate_model_ptr}
\title{Evaluate a model pointer at theta for testing the model}
\usage{
evaluate_model_ptr(theta, model_ptr)
}
\arguments{
\item{theta}{parameter vector}

\item{model_ptr}{external pointer to a ModelFn}
}
\description{
Evaluate a model pointer at theta for testing the model
}
