% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmcx-package.R
\docType{package}
\name{hmcx-package}
\alias{hmcx}
\alias{hmcx-package}
\title{hmcx: Collection of functions for HMC- and NUTS-sampling}
\description{
Provides functions implementing a HMC-sampler and a NUTS-Sampler in R and C++. 
             Define your own functions in R and C++ with CppAD to obtain Bayesian estimates 
             with HMC or NUTS
}
\keyword{internal}
