% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_args.R
\name{make_args}
\alias{make_args}
\title{Make a list of control arguments}
\usage{
make_args(
  biter = 1000,
  burnin = 500,
  badapt1 = 100,
  badapt2 = 100,
  max_depth = 10,
  max_depth_adapt = 10,
  epsilon = 0.1,
  L = 20,
  delta = 0.8,
  gamma = 0.05,
  t0 = 10,
  kappa = 0.75
)
}
\arguments{
\item{biter}{Length of the chain.}

\item{burnin}{Number of burnin samples}

\item{badapt1}{Number of samples to find optimal transition matrix, defaults to 100}

\item{max_depth}{Determines the number of maximal leapfrog steps in NUTS, defaults to 10, i.e. 2^10 = 1,024 steps}

\item{max_depth_adapt}{Similar to max_depth but for the warmup-phase, defaults to 10}

\item{epsilon}{A default stepsize, defaults to 0.10}

\item{L}{The number of steps in HMC, defaults to 20}

\item{...}{some further arguments for dual averaging}
}
\description{
Make a list of control arguments
}
