% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuts_cpp.R
\name{nuts_chain_cpp}
\alias{nuts_chain_cpp}
\title{Obtain samples from a NUTS-chain with all subfunctions implemented in C++}
\usage{
nuts_chain_cpp(
  model_ptr = NULL,
  args = NULL,
  verbose = NULL,
  inits = NULL,
  find_epsilon = TRUE
)
}
\arguments{
\item{model_ptr}{A pointer to the model to be estimated. Expects a single argument, the parameter vector.}

\item{args}{A list of control arguments, see make_args()}

\item{verbose}{A boolean argument to see the progress of the sampler, defaults to FALSE}

\item{inits}{A vector of initial parameter values to start the chain}

\item{find_epsilon}{A boolean argument to control whether an initial stepsize should be obtained with bracketing, defaults to TRUE}
}
\description{
Obtain samples from a NUTS-chain with all subfunctions implemented in C++
}
