% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuts.R
\name{nuts_chain_r}
\alias{nuts_chain_r}
\title{Obtain samples from a NUTS-chain with all subroutines implemented in R}
\usage{
nuts_chain_r(
  model_fn = NULL,
  args = NULL,
  verbose = FALSE,
  inits = NULL,
  find_epsilon = TRUE
)
}
\arguments{
\item{model_fn}{A closure for the model to be estimated. Expects a single argument, the parameter vector.}

\item{args}{A list of control arguments, see make_args()}

\item{verbose}{A boolean argument to see the progress of the sampler, defaults to FALSE}

\item{inits}{A vector of initial parameter values to start the chain}

\item{find_epsilon}{A boolean argument to control whether an initial stepsize should be obtained with bracketing, defaults to TRUE}
}
\description{
Obtain samples from a NUTS-chain with all subroutines implemented in R
}
